backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
google-genai==1.5.0
sqlalchemy==2.0.35
pydantic==2.9.0
python-multipart==0.0.12
PyPDF2==3.0.1
opencv-python-headless
numpy
pillow
python-dotenv==1.0.1
aiofiles==24.1.0
